requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.8.0
kiwisolver==1.1.0
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
Pillow==7.0.0
pyparsing==2.4.6
PySide2==5.14.1
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.4.1
shiboken2==5.14.1
six==1.14.0
